# docs/api/openapi-v1.0.0.yaml
openapi: 3.1.0
info:
  title: Tunnara Control API
  version: 1.0.0
  description: API de gestão, provisionamento, Cloudflare/DNS, nós distribuídos e redes privadas.
servers:
  - url: https://control.example.com
security:
  - bearerAuth: []
paths:
...
  /api/v1/agents/register:
    post:
      security: []
      parameters:
        - in: header
          name: X-Tunnara-Provisioning-Token
          required: true
          schema: {type: string}
      responses:
        '201': {description: Agent registrado}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...